menu fixes
