+++ app/layouts/fullWidth.vue
